move microservices to subfolder/module + fix test
