+++ src/scalecube-services/services/Message.js
