+++ src/scalecube-services/services/Router.js
@@ -1,5 +1,5 @@
 // @flow
-import { Message, ServiceInstance } from 'src/scalecube-services';
+import { Message, ServiceInstance } from 'src/scalecube-services/services';
 
 export interface Router {
   route(message:Message): ServiceInstance | null;
