+++ src/scalecube-services/services/ServiceInstance.js
