+++ src/scalecube-services/services/index.js
