+++ src/scalecube-services/services/libdef.js
